+++ firmware/src/system/mod.rs
@@ -1,19 +1,33 @@
-use daisy::hal;
+pub use daisy::hal;
+
+use daisy::led::LedUser;
 use hal::pac::CorePeripherals;
+use hal::pac::Peripherals as DevicePeripherals;
 use systick_monotonic::Systick;
 
 pub struct System {
     pub mono: Systick<1000>,
+    pub led_user: LedUser,
 }
 
 impl System {
+    /// Initialize system abstraction
+    ///
+    /// # Panics
+    ///
+    /// The system can be initialized only once. It panics otherwise.
     #[must_use]
-    pub fn init(mut cp: CorePeripherals) -> Self {
+    pub fn init(mut cp: CorePeripherals, dp: DevicePeripherals) -> Self {
         enable_cache(&mut cp);
 
-        let mono = Systick::new(cp.SYST, 480_000_000);
+        let board = daisy::Board::take().unwrap();
+        let ccdr = daisy::board_freeze_clocks!(board, dp);
+        let pins = daisy::board_split_gpios!(board, ccdr, dp);
 
-        Self { mono }
+        let mono = Systick::new(cp.SYST, 480_000_000);
+        let led_user = daisy::board_split_leds!(pins).USER;
+
+        Self { mono, led_user }
     }
 }